refactor(build): share the trim-internal-dts api-extractor plugin
Both packages' vite configs duplicated the ~80-line api-extractor trim pass that strips @internal members and the leftover 'Excluded from this release type' stubs from the rolled-up .d.ts. Extract it into scripts/vite-trim-internal-dts.ts as a parameterized plugin factory and have both configs import it.

Options cover the divergences: 'entries' (single index for babylon-lite, all rolled-up entries for compat), 'internalMissingUnderscore' (error vs off — compat allows @internal on the camera constructor overloads), 'silenceWrongInputFileType' (compat follows types into the babylon-lite peer's .ts source), and a 'transform' hook (compat rewrites the babylon-lite -> @babylonjs/lite import specifier). api-extractor is owned by the root devDependency that backs the shared scripts helper, so the redundant copy added to the compat package is dropped. Both builds produce byte-identical output to before; eslint, prettier, and the standalone typecheck pass.
