+++ packages/babylon-lite-compat/package.json
@@ -51,7 +51,6 @@
         "@recast-navigation/wasm": "0.43.0"
     },
     "devDependencies": {
-        "@microsoft/api-extractor": "^7.58.7",
         "@webgpu/types": "^0.1.52",
         "typescript": "^5.7.0",
         "vite": "^6.0.0",
